task2_animal-multimodal_classifier/pyproject.toml
[project]
name = "task2-animal-multimodal-classifier"
version = "0.1.0"
description = ""
authors = [
    {name = "Pasha0923",email = "alexsai@i.ua"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "numpy (>=2.4.3,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "transformers (>=5.3.0,<6.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "torchvision (>=0.25.0,<0.26.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "accelerate (>=1.13.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "notebook (>=7.5.4,<8.0.0)",
    "datasets (>=4.7.0,<5.0.0)"
]
